fix: address second review — stability formula, repr safety, dtype
- Simplify stability to (stab_down + stab_up) / 2, fixing double-count
  of symmetric FM scores for interior K values
- Use getattr in Label.__repr__ lambda for safety
- Cast pickle codes to np.intp for pandas 2.x Categorical compat
- Remove premature _stability backward-compat alias
- Clarify n_clusters docstring

+++ omicverse/io/general/_serialization.py
@@ -30,7 +30,7 @@ def _patch_categorical_setstate():
                 # Reconstruct via dict-based __setstate__ which pandas 2.x supports
                 dict_state = {
                     "_dtype": dtype,
-                    "_ndarray": np.asarray(codes),
+                    "_ndarray": np.asarray(codes, dtype=np.intp),
                 }
                 return _orig(self, dict_state)
         return _orig(self, state)
